+++ tests/conftest.py
@@ -41,3 +41,7 @@ def schema_samples_file_path(schemas_path):
 @pytest.fixture
 def schema_invalid_file_path(schemas_path):
     return os.path.join(schemas_path, "test_schema_invalid.yaml")
+
+@pytest.fixture
+def schema_sample_invalid_file_path(schemas_path):
+    return os.path.join(schemas_path, "test_schema_sample_invalid.yaml")